Optimize optimal parser memory: Reduce parent_len to uint16_t
The `parent_len` array, part of the optimal parser's scratch memory, previously used `uint32_t` to store match lengths. As typical match lengths do not exceed `UINT16_MAX`, this commit reduces its data type to `uint16_t`.

This change halves the memory consumed by `parent_len`, contributing to a further reduction in the overall `opt_scratch` region from approximately 10.125x `src_size` to 8.125x `src_size`.

All affected size estimations and documentation comments have been updated to reflect this memory reduction.

+++ src/lib/zxc_common.c
@@ -636,7 +636,7 @@ uint64_t zxc_decompress_block_bound(const size_t uncompressed_size) {
  * is rounded up to the cache-line boundary, so the returned value matches the
  * single aligned allocation performed by the initializer.
  *
- * For @p level >= 6 the figure also includes ctx->opt_scratch (~10.125 bytes
+ * For @p level >= 6 the figure also includes ctx->opt_scratch (~8.125 bytes
  * per chunk_size byte: dp + parent_len + parent_off + a 1-bit-per-position
  * match-end bitmap), the cache-line-aligned scratch used by the optimal
  * parser. It is lazy-allocated on the first level-6 call and persists for
@@ -666,7 +666,7 @@ uint64_t zxc_estimate_cctx_size(const size_t src_size, const int level) {
     if (level >= ZXC_LEVEL_DENSITY) {
         const size_t n_bm_words = (chunk_size + 1 + 63) / 64;
         total += ZXC_ALIGN_CL((chunk_size + 1) * sizeof(uint32_t)); /* dp             */
-        total += ZXC_ALIGN_CL((chunk_size + 1) * sizeof(uint32_t)); /* parent_len     */
+        total += ZXC_ALIGN_CL((chunk_size + 1) * sizeof(uint16_t)); /* parent_len     */
         total += ZXC_ALIGN_CL((chunk_size + 1) * sizeof(uint16_t)); /* parent_off     */
         total += ZXC_ALIGN_CL(n_bm_words * sizeof(uint64_t));       /* match_end_bits */
     }
